requirements.txt
gTTS==2.5.4
langchain_core==0.3.28
langchain_groq==0.2.2
pygame==2.6.1
PyPDF2==3.0.1
python-dotenv==1.0.1
SpeechRecognition==3.12.0
streamlit==1.41.1
